make a generic disk job type
