+++ src/disk_job.cpp
@@ -9,12 +9,7 @@ You may use, distribute and modify this code under the terms of the BSD license,
 see LICENSE file.
 */
 
-#include "libtorrent/aux_/mmap_disk_job.hpp"
-#include "libtorrent/disk_buffer_holder.hpp"
-
-#include "libtorrent/aux_/disable_warnings_push.hpp"
-#include <boost/variant/get.hpp>
-#include "libtorrent/aux_/disable_warnings_pop.hpp"
+#include "libtorrent/aux_/disk_job.hpp"
 
 namespace libtorrent {
 namespace aux {
@@ -22,7 +17,7 @@ namespace aux {
 namespace {
 	struct caller_visitor
 	{
-		explicit caller_visitor(mmap_disk_job& j) : m_job(j) {}
+		explicit caller_visitor(disk_job& j) : m_job(j) {}
 
 		void operator()(job::read& j) const
 		{
@@ -103,10 +98,10 @@ namespace {
 		}
 
 	private:
-		mmap_disk_job& m_job;
+		disk_job& m_job;
 	};
 }
-	void mmap_disk_job::call_callback()
+	void disk_job::call_callback()
 	{
 		std::visit(caller_visitor(*this), action);
 	}
