+++ src/disk_job_fence.cpp
@@ -11,7 +11,7 @@ see LICENSE file.
 
 
 #include "libtorrent/aux_/disk_job_fence.hpp"
-#include "libtorrent/aux_/mmap_disk_job.hpp"
+#include "libtorrent/aux_/disk_job.hpp"
 #include "libtorrent/performance_counters.hpp"
 
 #define DEBUG_STORAGE 0
@@ -25,16 +25,16 @@ see LICENSE file.
 namespace libtorrent {
 namespace aux {
 
-	int disk_job_fence::job_complete(mmap_disk_job* j, tailqueue<mmap_disk_job>& jobs)
+	int disk_job_fence::job_complete(disk_job* j, tailqueue<disk_job>& jobs)
 	{
 		std::lock_guard<std::mutex> l(m_mutex);
 
-		TORRENT_ASSERT(j->flags & mmap_disk_job::in_progress);
-		j->flags &= ~mmap_disk_job::in_progress;
+		TORRENT_ASSERT(j->flags & disk_job::in_progress);
+		j->flags &= ~disk_job::in_progress;
 
 		TORRENT_ASSERT(m_outstanding_jobs > 0);
 		--m_outstanding_jobs;
-		if (j->flags & mmap_disk_job::fence)
+		if (j->flags & disk_job::fence)
 		{
 			// a fence job just completed. Make sure the fence logic
 			// works by asserting m_outstanding_jobs is in fact 0 now
@@ -49,8 +49,8 @@ namespace aux {
 			int ret = 0;
 			while (!m_blocked_jobs.empty())
 			{
-				mmap_disk_job *bj = m_blocked_jobs.pop_front();
-				if (bj->flags & mmap_disk_job::fence)
+				disk_job *bj = m_blocked_jobs.pop_front();
+				if (bj->flags & disk_job::fence)
 				{
 					// we encountered another fence. We cannot post anymore
 					// jobs from the blocked jobs queue. We have to go back
@@ -59,8 +59,8 @@ namespace aux {
 					// executing currently, we should add the fence job.
 					if (m_outstanding_jobs == 0 && jobs.empty())
 					{
-						TORRENT_ASSERT(!(bj->flags & mmap_disk_job::in_progress));
-						bj->flags |= mmap_disk_job::in_progress;
+						TORRENT_ASSERT(!(bj->flags & disk_job::in_progress));
+						bj->flags |= disk_job::in_progress;
 						++m_outstanding_jobs;
 						++ret;
 #if TORRENT_USE_ASSERTS
@@ -76,8 +76,8 @@ namespace aux {
 					}
 					return ret;
 				}
-				TORRENT_ASSERT(!(bj->flags & mmap_disk_job::in_progress));
-				bj->flags |= mmap_disk_job::in_progress;
+				TORRENT_ASSERT(!(bj->flags & disk_job::in_progress));
+				bj->flags |= disk_job::in_progress;
 
 				++m_outstanding_jobs;
 				++ret;
@@ -100,11 +100,11 @@ namespace aux {
 		TORRENT_ASSERT(m_blocked_jobs.size() > 0);
 
 		// this is the fence job
-		mmap_disk_job *bj = m_blocked_jobs.pop_front();
-		TORRENT_ASSERT(bj->flags & mmap_disk_job::fence);
+		disk_job *bj = m_blocked_jobs.pop_front();
+		TORRENT_ASSERT(bj->flags & disk_job::fence);
 
-		TORRENT_ASSERT(!(bj->flags & mmap_disk_job::in_progress));
-		bj->flags |= mmap_disk_job::in_progress;
+		TORRENT_ASSERT(!(bj->flags & disk_job::in_progress));
+		bj->flags |= disk_job::in_progress;
 
 		++m_outstanding_jobs;
 #if TORRENT_USE_ASSERTS
@@ -116,7 +116,7 @@ namespace aux {
 		return 1;
 	}
 
-	bool disk_job_fence::is_blocked(mmap_disk_job* j)
+	bool disk_job_fence::is_blocked(disk_job* j)
 	{
 		std::lock_guard<std::mutex> l(m_mutex);
 		DLOG(stderr, "[%p] is_blocked: fence: %d num_outstanding: %d\n"
@@ -127,8 +127,8 @@ namespace aux {
 		// this job still needs to get queued up
 		if (m_has_fence == 0)
 		{
-			TORRENT_ASSERT(!(j->flags & mmap_disk_job::in_progress));
-			j->flags |= mmap_disk_job::in_progress;
+			TORRENT_ASSERT(!(j->flags & disk_job::in_progress));
+			j->flags |= disk_job::in_progress;
 			++m_outstanding_jobs;
 			return false;
 		}
@@ -156,11 +156,11 @@ namespace aux {
 	}
 
 	// j is the fence job. It must have exclusive access to the storage
-	int disk_job_fence::raise_fence(mmap_disk_job* j, counters& cnt)
+	int disk_job_fence::raise_fence(disk_job* j, counters& cnt)
 	{
-		TORRENT_ASSERT(!(j->flags & mmap_disk_job::in_progress));
-		TORRENT_ASSERT(!(j->flags & mmap_disk_job::fence));
-		j->flags |= mmap_disk_job::fence;
+		TORRENT_ASSERT(!(j->flags & disk_job::in_progress));
+		TORRENT_ASSERT(!(j->flags & disk_job::fence));
+		j->flags |= disk_job::fence;
 
 		std::lock_guard<std::mutex> l(m_mutex);
 
@@ -177,7 +177,7 @@ namespace aux {
 			// after this, without being passed through is_blocked()
 			// that's why we're accounting for it here
 
-			j->flags |= mmap_disk_job::in_progress;
+			j->flags |= disk_job::in_progress;
 			++m_outstanding_jobs;
 			return fence_post_fence;
 		}
