+++ test/test_fence.cpp
@@ -10,7 +10,7 @@ You may use, distribute and modify this code under the terms of the BSD license,
 see LICENSE file.
 */
 
-#include "libtorrent/aux_/mmap_disk_job.hpp"
+#include "libtorrent/aux_/disk_job.hpp"
 #include "libtorrent/aux_/disk_job_fence.hpp"
 #include "libtorrent/performance_counters.hpp"
 #include "test.hpp"
@@ -20,14 +20,14 @@ see LICENSE file.
 using namespace lt;
 
 using lt::aux::disk_job_fence;
-using lt::aux::mmap_disk_job;
+using lt::aux::disk_job;
 
 TORRENT_TEST(empty_fence)
 {
 	disk_job_fence fence;
 	counters cnt;
 
-	mmap_disk_job test_job[10];
+	disk_job test_job[10];
 
 	// issue 5 jobs. None of them should be blocked by a fence
 	int ret_int = 0;
@@ -43,7 +43,7 @@ TORRENT_TEST(empty_fence)
 	ret = fence.is_blocked(&test_job[8]);
 	TEST_CHECK(ret == true);
 
-	aux::tailqueue<aux::mmap_disk_job> jobs;
+	aux::tailqueue<aux::disk_job> jobs;
 
 	// complete the fence job
 	fence.job_complete(&test_job[5], jobs);
@@ -64,7 +64,7 @@ TORRENT_TEST(job_fence)
 	counters cnt;
 	disk_job_fence fence;
 
-	mmap_disk_job test_job[10];
+	disk_job test_job[10];
 
 	// issue 5 jobs. None of them should be blocked by a fence
 	int ret_int = 0;
@@ -96,7 +96,7 @@ TORRENT_TEST(job_fence)
 	ret = fence.is_blocked(&test_job[8]);
 	TEST_CHECK(ret == true);
 
-	aux::tailqueue<aux::mmap_disk_job> jobs;
+	aux::tailqueue<aux::disk_job> jobs;
 
 	fence.job_complete(&test_job[3], jobs);
 	TEST_CHECK(jobs.size() == 0);
@@ -134,7 +134,7 @@ TORRENT_TEST(double_fence)
 	counters cnt;
 	disk_job_fence fence;
 
-	mmap_disk_job test_job[10];
+	disk_job test_job[10];
 
 	// issue 5 jobs. None of them should be blocked by a fence
 	int ret_int = 0;
@@ -170,7 +170,7 @@ TORRENT_TEST(double_fence)
 	ret = fence.is_blocked(&test_job[9]);
 	TEST_CHECK(ret == true);
 
-	aux::tailqueue<aux::mmap_disk_job> jobs;
+	aux::tailqueue<aux::disk_job> jobs;
 
 	fence.job_complete(&test_job[3], jobs);
 	TEST_CHECK(jobs.size() == 0);
